requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.1
aiosignal==1.3.1
alembic==1.16.4
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.3.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.14
Django==5.0.6
effdet==0.4.1
emoji==2.14.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
executing==2.1.0
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.21.0
google-api-python-client==2.176.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.7.4
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
iopath==0.1.10
IProgress==0.4
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==3.7.0
kiwisolver==1.4.7
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.11
langchain-text-splitters==0.3.0
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.1.135
language_data==1.3.0
layoutparser==0.3.4
libclang==18.1.1
lxml==5.3.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.13
mypy-extensions==1.0.0
mysql-connector-python==9.3.0
namex==0.0.8
narwhals==1.46.0
nest-asyncio==1.6.0
networkx==3.4.1
nltk==3.9.1
numpy==2.1.3
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.18.0
pikepdf==9.3.0
pillow==11.0.0
platformdirs==4.3.6
portalocker==2.10.1
preshed==3.0.10
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.24.11
pypandoc==1.14
pyparsing==3.2.0
pypdf==5.0.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.12
python-oxmsg==0.0.1
python-pptx==1.0.2
python-telegram-bot==22.2
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9
safetensors==0.4.5
schedule==1.2.2
scipy==1.13.1
sentencepiece==0.2.0
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
sqlparse==0.5.0
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.46.1
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.19.0
thinc==8.3.6
timm==1.0.10
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torchaudio==2.4.1+cu118
torchvision==0.22.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.53.2
twilio==9.6.5
typer==0.16.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.149
ultralytics-thop==2.0.14
unstructured==0.15.14
unstructured-client==0.26.1
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.13
uritemplate==4.2.0
urllib3==2.2.3
uvicorn==0.34.3
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.15.4
